added newline for clang, changed int64_t to yasl_int in parser ast and env

+++ compiler/compiler/env/env.c
@@ -29,12 +29,12 @@ void env_del_current_only(Env_t *env) {
     free(env);
 }
 
-int64_t env_len(Env_t *env) {
+size_t env_len(Env_t *env) {
     if (env == NULL) return 0;
     return env->vars->count + env_len(env->parent);
 }
 
-int env_contains_cur_scope(Env_t *env, char *name, int64_t name_len) {
+int env_contains_cur_scope(Env_t *env, char *name, size_t name_len) {
     String_t *string = str_new_sized_heap(0, name_len, copy_char_buffer(name_len, name));
     struct YASL_Object key = YASL_STR(string); // (struct YASL_Object) { .value.sval = string, .type = Y_STR };
 
@@ -46,7 +46,7 @@ int env_contains_cur_scope(Env_t *env, char *name, int64_t name_len) {
     return 1;
 }
 
-int env_contains(Env_t *env, char *name, int64_t name_len) {
+int env_contains(Env_t *env, char *name, size_t name_len) {
     if (env == NULL) return 0;
     String_t *string = str_new_sized_heap(0, name_len, copy_char_buffer(name_len, name));
     struct YASL_Object key = YASL_STR(string);
@@ -60,7 +60,7 @@ int env_contains(Env_t *env, char *name, int64_t name_len) {
     return 1;
 }
 
-int64_t env_get(Env_t *env, char *name, int64_t name_len) {
+int64_t env_get(Env_t *env, char *name, size_t name_len) {
     String_t *string = str_new_sized_heap(0, name_len, copy_char_buffer(name_len, name));
     struct YASL_Object key = YASL_STR(string);
 
@@ -75,19 +75,19 @@ int64_t env_get(Env_t *env, char *name, int64_t name_len) {
     return value.value.ival;
 }
 
-void env_decl_var(Env_t *env, char *name, int64_t name_len) {
+void env_decl_var(Env_t *env, char *name, size_t name_len) {
     String_t *string = str_new_sized_heap(0, name_len, copy_char_buffer(name_len, name));
     struct YASL_Object key = YASL_STR(string);
     struct YASL_Object value = YASL_INT(env_len(env));
     table_insert(env->vars, key, value);
 }
 
-static struct Table *get_closest_scope_with_var(Env_t *env, char *name, int64_t name_len) {
+static struct Table *get_closest_scope_with_var(Env_t *env, char *name, size_t name_len) {
     struct YASL_Object key = table_search_string_int(env->vars, name, name_len);
     return key.type != Y_END ? env->vars : get_closest_scope_with_var(env->parent, name, name_len);
 }
 
-void env_make_const(Env_t *env, char *name, int64_t name_len) {
+void env_make_const(Env_t *env, char *name, size_t name_len) {
     struct Table *ht = get_closest_scope_with_var(env, name, name_len);
     table_insert_string_int(ht, name, name_len, ~table_search_string_int(ht, name, name_len).value.ival);
 }
